Move lending period methods to concerns
